refactor webAudioUnlock
